Time: 4 ms (58.51%), Space: 16.6 MB (22.6%) - LeetHub

+++ 1050-construct-binary-search-tree-from-preorder-traversal/1050-construct-binary-search-tree-from-preorder-traversal.cpp
@@ -11,20 +11,20 @@
  */
 class Solution {
 public:
-    TreeNode *solve(int & index, vector<int> & pre , int mini , int maxi ){
+    TreeNode * solve(int & index, vector<int> & pre , long mini , long maxi){
         if(index >= pre.size()) return nullptr;
 
-        if(pre[index] >=maxi || pre[index] <= mini) return nullptr;
+        if(pre[index] >= maxi || pre[index] <= mini) return nullptr;
 
-        TreeNode * root = new TreeNode(pre[index++]);
-        root->left = solve(index,pre,mini , root->val);
-        root->right = solve(index,pre,root->val,maxi);
+        TreeNode * root =new TreeNode(pre[index++]);
+
+        root->left = solve(index,pre,mini,root->val);
+        root->right = solve(index,pre,root->val ,maxi);
 
         return root;
     }
     TreeNode* bstFromPreorder(vector<int>& preorder) {
-       int index = 0;
-
-       return solve(index,preorder,INT_MIN, INT_MAX);  
+        int index = 0;
+        return solve(index, preorder, LONG_MIN , LONG_MAX );
     }
 };
